Updated and properly added the Level 2 code...
